// src/modules/v1/v1.interface.ts
export interface CardDetail {
  id: string;
  assetCommonName: string;
  asset: string;
  assetLongname: string;
  assetgroup: string;
  cardName: string;
  ownerName: string;
  imgurUrl: string;
  addDescription: string;
  twId: string;
  twName: string;
  tag: string;
  cid: string;
  ver: number;
  isGoodStatus: boolean;
  registTime: number;
  updateTime: number;
}